+++ index.es6.js
@@ -1,3 +1,4 @@
 export { default as default } from './source/form'
 export { default as Field }   from './source/field'
+export { default as Submit }  from './source/submit'
 export { default as reducer } from './source/reducer'