+++ test/exports.js
@@ -3,6 +3,7 @@ import React from 'react'
 import Form,
 {
 	Field,
+	Submit,
 	reducer
 }
 from '../index.es6'
@@ -15,6 +16,7 @@ describe(`exports`, function()
 		const render = <Decorated_form/>
 		
 		const field_render = <Field name="test" component={props => <div>Test</div>}/>
+		const submit_render = <Submit component={props => <button>Test</button>}/>
 
 		reducer({}, {})
 	})
@@ -28,6 +30,9 @@ describe(`exports`, function()
 		
 		const Field = Form.Field
 		const field_render = <Field name="test" component={props => <div>Test</div>}/>
+
+		const Submit = Form.Submit
+		const submit_render = <Submit component={props => <button>Test</button>}/>
 
 		Form.reducer({}, {})
 	})